requirements.txt
algorithmia>=1.0.0,<2.0
torchvision
pillow